requirements.txt
numpy~=2.4.2
pandas~=3.0.0
mmead~=0.1.0
tqdm~=4.67.3
click~=8.3.1
matplotlib~=3.10.8
scipy~=1.17.0
pytrec-eval-terrier